+++ analytics/urls.py
@@ -4,11 +4,9 @@ from django.conf.urls import include
 from django.urls import path
 from django.urls.resolvers import URLPattern, URLResolver
 
-from analytics.views.installation_activity import (
-    get_installation_activity,
-    get_remote_server_activity,
-)
+from analytics.views.installation_activity import get_installation_activity
 from analytics.views.realm_activity import get_realm_activity
+from analytics.views.remote_activity import get_remote_server_activity
 from analytics.views.stats import (
     get_chart_data,
     get_chart_data_for_installation,
